Web API
* Added Last Reading Static Field

Placeholder for a future Database, this static field just stores the last value for ease of retrieval.

* Added a test

The test verifies the API is returning a response code 200 (successful) and prints out the request for manual verification rn

* Create WebController.java

Controller for the API, doesn't work by itself

* Created WebRepository

Moved the storage of Sensor Data from the Sensor
Reading class to Web Repository. Although static,
easily converted into an actual Database Controller
later

* Modified the response type and content

Now the response is a manually generated JSON
string that contains an array of, for now, all readings.
Now the test passes and all responses are
answered, even with an empty array before any
readings are made, which is why it wasn't working
before

* Changed name of the endpoint

The endpoint /lastReading was returning all readings,
so changed to /readings to be more descriptive

* Made the requests null-safe

I tried my best to make the requests null-safe, so that
even when provided with null SensorReadings it can
still manage to respond

* Added another endpoint

Since we might just want to get the last reading

* Updated Web-API test

Abstracted the tests so that they can test any
endpoint easily without code repetition. At least
simple tests