Revamped Lorawan module
Strap in for a long commit message, I have tried to better separate concerns among all of the classes included in the Lorawan module. They've also been all moved for the same reason the data classes were, Spring auto-discovery

As such, the client has acquired the responsibility of implementing the LorawanEventHandler interface that provides a method to handle each command defined by the Lorawan server.

It is also responsible for propagating the necessary objects to the Listener/Dispatcher/PayloadHandler that it instantiates. This is due to the behavior of @Autowire and the highly specific order that these elements need to be instantiated in to ensure that they are available to instantiate the following objects.

The responsibility for taking the steps towards initialization have moved over to the Lorawan application class instead.

The payload handler has been renamed data handler, as its new responsibility is to mostly handle data processing by parsing payloads to and from the database and ensuring the data is correctly formatted

The listener responsibility is to listen for all messages received from the server and delegate their handling to the client or logging events happening in the Websocket communication

Finally, the SocketDispatcher is responsible for sending messages to the server, and when combined with the listener, allow the client to communicate bidirectionally with the websocket.