Websocket Keep-Alive
This commit introduces a terrible bug, the application infinitely increases a counter every time it is queried for a refresh, unless the socket is closed. Assuming the socket doesn't close, this means the application will inevitably overflow after roughly 3 quadrillion years of continuous runtime, assuming a 3 minute probe time, and I don't think we should care.